Colour changing sky depending on daytime
A sky shader with colours consistent with daytime, and stars at night.
Source: https://github.com/BastiaanOlij/godot-sky-asset
The shader is modified and integrated into the scene as a Skybox.
Also added a slider to the Pause menu to change time during game play,
and removed the obsolete Level Select menu.